+++ src/pages/MainPage.tsx
@@ -3,7 +3,7 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 import { menuSelectionState } from '../recoil/apiDataAtoms';
 import Menu from '../components/Menu/Menu';
 import RegionSelection from '../components/Selection/RegionSelection';
-import SelectionResult from '../components/Selection/SelectionResult';
+// import SelectionResult from '../components/Selection/SelectionResult';
 import SliderBanner from '../components/SliderBanner';
 // import StayRecommendation from '../components/Recommendation/StayRecommendation';
 // import RestaurantRecommendation from '../components/Recommendation/RestaurantRecommendation';
@@ -15,6 +15,7 @@ import BestDate from '../components/Recommendation/BestDate';
 import MySpot from '../components/Recommendation/MySpot';
 import { useEffect } from 'react';
 import WeeklyTop10v2 from '../components/Recommendation/WeeklyTop10v2';
+import SpotSelectionResult from '../components/Selection/SpotSelectionResult';
 
 const MainPage = () => {
   const [selectedMenu, setSelectedMenu] = useRecoilState(menuSelectionState);
@@ -47,8 +48,9 @@ const MainPage = () => {
           <RestaurantRecommendation /> */}
         </>
       ) : selectedMenu === '관광지' ? (
-        <SelectionResult />
-      ) : selectedMenu === '숙박' ? (
+        <SpotSelectionResult />
+      ) : // <SelectionResult />
+      selectedMenu === '숙박' ? (
         <StaySelectionResult />
       ) : selectedMenu === '음식점' ? (
         <RestaurantSelectionResult />
